Fixed issue with code view when running context command.
Previously, using the context command would cause llef to render the
disassembly without the correct current instruction. This
is because the code looked for the presense of an arrow marker
to indicate the current instruction, which is not outputted
by LLDB when refreshing the context. To fix this, this
patch pulls the value of the pc and compares this
instead